2D: Addition of set/get render layers to handle object visibility. Canvas: For now canvas object will be relative also to 2d objects. ProgressBar has can be drawed as a 2d object now, this is temporary. Object: set_mark was removed. Physics: Disable vtable for drivers, P2D will be the unique implementation for now. Scene: SceneManager will no longer call current_scene->event twice.

+++ src/collections/string_utility.h
@@ -46,7 +46,7 @@ struct StringUtility
 			{
 				fractional_part *= 10.0f;
 				i32 digit = i32(fractional_part);
-				*it++ = '0' + digit;
+				*it++ = '0' + i8(digit);
 				fractional_part -= digit;
 			}
 		}
